LayoutTests/imported/w3c/web-platform-tests/css/css-syntax/anb-parsing-expected.txt


FAIL "odd" becomes "2n+1" assert_equals: expected "2n+1" but got "odd"
FAIL "even" becomes "2n" assert_equals: expected "2n" but got "even"
PASS "1" becomes "1" 
FAIL "+1" becomes "1" assert_equals: expected "1" but got "+1"
PASS "-1" becomes "-1" 
PASS "5n" becomes "5n" 
FAIL "5N" becomes "5n" assert_equals: expected "5n" but got "5N"
FAIL "+n" becomes "n" assert_equals: expected "n" but got "+n"
PASS "n" becomes "n" 
FAIL "N" becomes "n" assert_equals: expected "n" but got "N"
PASS "+ n" becomes "parse error" 
PASS "-n" becomes "-n" 
FAIL "-N" becomes "-n" assert_equals: expected "-n" but got "-N"
PASS "5n-5" becomes "5n-5" 
FAIL "+n-5" becomes "n-5" assert_equals: expected "n-5" but got "+n-5"
PASS "n-5" becomes "n-5" 
PASS "+ n-5" becomes "parse error" 
PASS "-n-5" becomes "-n-5" 
FAIL "5n +5" becomes "5n+5" assert_equals: expected "5n+5" but got "5n +5"
FAIL "5n -5" becomes "5n-5" assert_equals: expected "5n-5" but got "5n -5"
FAIL "+n +5" becomes "n+5" assert_equals: expected "n+5" but got "+n +5"
FAIL "n +5" becomes "n+5" assert_equals: expected "n+5" but got "n +5"
FAIL "+n -5" becomes "n-5" assert_equals: expected "n-5" but got "+n -5"
PASS "+ n +5" becomes "parse error" 
PASS "n 5" becomes "parse error" 
FAIL "-n +5" becomes "-n+5" assert_equals: expected "-n+5" but got "-n +5"
FAIL "-n -5" becomes "-n-5" assert_equals: expected "-n-5" but got "-n -5"
PASS "-n 5" becomes "parse error" 
FAIL "5n- 5" becomes "5n-5" assert_equals: expected "5n-5" but got "5n- 5"
PASS "5n- -5" becomes "parse error" 
PASS "5n- +5" becomes "parse error" 
FAIL "-5n- 5" becomes "-5n-5" assert_equals: expected "-5n-5" but got "-5n- 5"
FAIL "+n- 5" becomes "n-5" assert_equals: expected "n-5" but got "+n- 5"
FAIL "n- 5" becomes "n-5" assert_equals: expected "n-5" but got "n- 5"
PASS "+ n- 5" becomes "parse error" 
PASS "n- +5" becomes "parse error" 
PASS "n- -5" becomes "parse error" 
FAIL "-n- 5" becomes "-n-5" assert_equals: expected "-n-5" but got "-n- 5"
PASS "-n- +5" becomes "parse error" 
PASS "-n- -5" becomes "parse error" 
FAIL "5n + 5" becomes "5n+5" assert_equals: expected "5n+5" but got "5n + 5"
FAIL "5n - 5" becomes "5n-5" assert_equals: expected "5n-5" but got "5n - 5"
PASS "5n + +5" becomes "parse error" 
PASS "5n + -5" becomes "parse error" 
PASS "5n - +5" becomes "parse error" 
PASS "5n - -5" becomes "parse error" 
FAIL "+n + 5" becomes "n+5" assert_equals: expected "n+5" but got "+n + 5"
FAIL "n + 5" becomes "n+5" assert_equals: expected "n+5" but got "n + 5"
PASS "+ n + 5" becomes "parse error" 
FAIL "+n - 5" becomes "n-5" assert_equals: expected "n-5" but got "+n - 5"
PASS "+n + +5" becomes "parse error" 
PASS "+n + -5" becomes "parse error" 
PASS "+n - +5" becomes "parse error" 
PASS "+n - -5" becomes "parse error" 
FAIL "-n + 5" becomes "-n+5" assert_equals: expected "-n+5" but got "-n + 5"
FAIL "-n - 5" becomes "-n-5" assert_equals: expected "-n-5" but got "-n - 5"
PASS "-n + +5" becomes "parse error" 
PASS "-n + -5" becomes "parse error" 
PASS "-n - +5" becomes "parse error" 
PASS "-n - -5" becomes "parse error" 
PASS "1 - n" becomes "parse error" 
PASS "0 - n" becomes "parse error" 
PASS "-1 + n" becomes "parse error" 
PASS "2 n + 2" becomes "parse error" 
PASS "- 2n" becomes "parse error" 
PASS "+ 2n" becomes "parse error" 
PASS "+2 n" becomes "parse error" 
